Updated and renamed some scripts
More work to be done:
01-combine_csv: sapply for 25 degrees section has stopped working. Need
to figure this out, but not highest priority as have combined data and
need to work on analysis.

03-visualisation_analysis - more viz to be done and graphing to come.

04-glm_analysis - working on glm analysis now.

CHANGES - tried to change to relative directories rather than absolute.

TODO - 01-combine: not working for 25degrees. Need to figure this out.

01-combine: need to figure way of setting wd individually and then move
back to home directory for saving. THIS IS NOT ELEGANT but is quick. I
will revist to redo.